+++ es40/src/Cirrus.cpp
@@ -29,6 +29,9 @@
  *
  * $Id$
  *
+ * X-1.14       Brian Wheeler                                   27-FEB-2008
+ *      Avoid compiler warnings.
+ *
  * X-1.13       David Leonard                                   20-FEB-2008
  *      Shut down refresh thread when emulator exits.
  *
@@ -524,7 +527,7 @@ int CCirrus::SaveState(FILE *f)
   fwrite(&ss,sizeof(long),1,f);
   fwrite(&state,sizeof(state),1,f);
   fwrite(&cirrus_magic2,sizeof(u32),1,f);
-  printf("%s: %d bytes saved.\n",devid_string,ss);
+  printf("%s: %d bytes saved.\n",devid_string,(int)ss);
   return 0;
 }
 
@@ -586,7 +589,7 @@ int CCirrus::RestoreState(FILE *f)
     return -1;
   }
 
-  printf("%s: %d bytes restored.\n",devid_string,ss);
+  printf("%s: %d bytes restored.\n",devid_string,(int)ss);
   return 0;
 }
